+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.0.11
+
+- fixed: icon not shown in bookmark view on Windows
+
 ## 0.0.10
 
 - fixed: errors due to assignment to const